Implemented Delete logic and also printed success message

+++ src/main/webapp/student-list.jsp
@@ -8,6 +8,15 @@
         <h2>MVC CRUD APPLICATION</h2>
         <h5>SERVLET + JSP + JDBC</h5>
 
+        <%
+            String success = (String)request.getParameter("success");
+            if(success != null){
+        %>
+            <p style='color:green;'><%=success%></p>
+        <%
+            }
+        %>
+
         <a href="students?action=add">Add Student</a>
         <table border="1" cellpadding="8">
             <thead>
@@ -24,7 +33,7 @@
             <%
                 int cnt = 1;
                 List<Student> students = (List<Student>)request.getAttribute("students");
-                if(students != null || !students.isEmpty()){
+                if(students != null && !students.isEmpty()){
                     for(Student s : students){
             %>
                         <tr>
@@ -34,7 +43,7 @@
                             <td><%= s.getMobile() %></td>
                             <td>
                                 <a href="students?action=edit&id=<%= s.getId() %>">Update</a>
-                                <a href="students?action=delete&id=<%= s.getId() %>">Delete</a>
+                                <a href="students?action=delete&id=<%= s.getId() %>" onclick="return confirm('Are you sure to delete?')">Delete</a>
                             </td>
                         </tr>
             <%
@@ -42,7 +51,7 @@
                 } else {
             %>
                     <tr>
-                        <td>No Students Found</t>
+                        <td colspan="5">No Students Found</t>
                     </tr>
             <%
                 }
